Build with fast math

+++ logs/log.cu
@@ -1,6 +1,6 @@
 #include <cmath>
 
-__global__ void mylog(double* value)
+__global__ void mylog(float* value)
 {
     value[threadIdx.x] = std::log(value[threadIdx.x]);
 }
